+++ docs/CHANGELOG.md
@@ -6,7 +6,7 @@
 
 - feat: add json schema validation for api
 - feat: add json schema validation for testcase v1 & v2
-- feat: add json schema validation for testsuite v2
+- feat: add json schema validation for testsuite v1 & v2
 
 **Changed**